MLP classifier example
