--- mdf_to_pytorch/example.py
@@ -1,6 +1,0 @@
-import torch
-import torch.nn
-import mdf2torch
-
-models = mdf2torch.load("examples/linear_test.json", eval_models=True)
-model = models["linear_test"]
